refactor(editor): cleanup dead code (#12281)
