--- blocksuite/affine/blocks/attachment/src/attachment-spec.ts
@@ -1,33 +1,0 @@
-import { AttachmentBlockSchema } from '@blocksuite/affine-model';
-import { SlashMenuConfigExtension } from '@blocksuite/affine-widget-slash-menu';
-import { BlockViewExtension, FlavourExtension } from '@blocksuite/std';
-import type { ExtensionType } from '@blocksuite/store';
-import { literal } from 'lit/static-html.js';
-
-import { AttachmentBlockAdapterExtensions } from './adapters/extension.js';
-import { AttachmentBlockInteraction } from './attachment-edgeless-block.js';
-import { AttachmentDropOption } from './attachment-service.js';
-import { attachmentSlashMenuConfig } from './configs/slash-menu.js';
-import { createBuiltinToolbarConfigExtension } from './configs/toolbar';
-import {
-  AttachmentEmbedConfigExtension,
-  AttachmentEmbedService,
-} from './embed';
-
-const flavour = AttachmentBlockSchema.model.flavour;
-
-export const AttachmentBlockSpec: ExtensionType[] = [
-  FlavourExtension(flavour),
-  BlockViewExtension(flavour, model => {
-    return model.parent?.flavour === 'affine:surface'
-      ? literal`affine-edgeless-attachment`
-      : literal`affine-attachment`;
-  }),
-  AttachmentDropOption,
-  AttachmentEmbedConfigExtension(),
-  AttachmentEmbedService,
-  AttachmentBlockAdapterExtensions,
-  AttachmentBlockInteraction,
-  createBuiltinToolbarConfigExtension(flavour),
-  SlashMenuConfigExtension(flavour, attachmentSlashMenuConfig),
-].flat();
