+++ blocksuite/affine/blocks/attachment/src/index.ts
@@ -1,7 +1,6 @@
 export * from './adapters';
 export * from './attachment-block';
 export * from './attachment-service';
-export * from './attachment-spec';
 export { attachmentViewDropdownMenu } from './configs/toolbar';
 export * from './edgeless-clipboard-config';
 export {
