+++ blocksuite/affine/blocks/bookmark/src/index.ts
@@ -1,6 +1,5 @@
 export * from './adapters';
 export * from './bookmark-block';
-export * from './bookmark-spec';
 export * from './commands';
 export * from './components';
 export { BookmarkSlashMenuConfigIdentifier } from './configs/slash-menu';
