+++ blocksuite/affine/blocks/callout/src/index.ts
@@ -1,3 +1,2 @@
 export * from './callout-block.js';
-export * from './callout-spec.js';
 export * from './effects.js';
