+++ blocksuite/affine/blocks/code/src/index.ts
@@ -2,7 +2,6 @@ export * from './adapters';
 export * from './clipboard';
 export * from './code-block';
 export * from './code-block-config';
-export * from './code-block-spec';
 export * from './code-preview-extension';
 export * from './code-toolbar';
 export * from './turbo/code-layout-handler';
