--- blocksuite/affine/blocks/divider/src/divider-spec.ts
@@ -1,10 +1,0 @@
-import { BlockViewExtension } from '@blocksuite/std';
-import type { ExtensionType } from '@blocksuite/store';
-import { literal } from 'lit/static-html.js';
-
-import { DividerBlockAdapterExtensions } from './adapters/extension.js';
-
-export const DividerBlockSpec: ExtensionType[] = [
-  BlockViewExtension('affine:divider', literal`affine-divider`),
-  DividerBlockAdapterExtensions,
-].flat();
