+++ blocksuite/affine/blocks/divider/src/index.ts
@@ -1,3 +1,2 @@
 export * from './adapters';
 export * from './divider-block';
-export * from './divider-spec';
