--- blocksuite/affine/blocks/edgeless-text/src/edgeless-text-spec.ts
@@ -1,10 +1,0 @@
-import { BlockViewExtension } from '@blocksuite/std';
-import type { ExtensionType } from '@blocksuite/store';
-import { literal } from 'lit/static-html.js';
-
-import { EdgelessTextInteraction } from './edgeless-text-block';
-
-export const EdgelessTextBlockSpec: ExtensionType[] = [
-  BlockViewExtension('affine:edgeless-text', literal`affine-edgeless-text`),
-  EdgelessTextInteraction,
-];
