+++ blocksuite/affine/blocks/edgeless-text/src/index.ts
@@ -1,4 +1,3 @@
 export * from './edgeless-clipboard-config';
 export * from './edgeless-text-block.js';
-export * from './edgeless-text-spec.js';
 export * from './edgeless-toolbar';
