+++ blocksuite/affine/blocks/embed-doc/src/embed-linked-doc-block/embed-linked-doc-spec.ts
@@ -3,23 +3,11 @@ import { BlockViewExtension } from '@blocksuite/std';
 import type { ExtensionType } from '@blocksuite/store';
 import { literal } from 'lit/static-html.js';
 
-import { EmbedLinkedDocBlockAdapterExtensions } from './adapters/extension';
 import { LinkedDocSlashMenuConfigExtension } from './configs/slash-menu';
 import { createBuiltinToolbarConfigExtension } from './configs/toolbar';
 import { EmbedLinkedDocInteraction } from './embed-edgeless-linked-doc-block';
 
 const flavour = EmbedLinkedDocBlockSchema.model.flavour;
-
-export const EmbedLinkedDocBlockSpec: ExtensionType[] = [
-  BlockViewExtension(flavour, model => {
-    return model.parent?.flavour === 'affine:surface'
-      ? literal`affine-embed-edgeless-linked-doc-block`
-      : literal`affine-embed-linked-doc-block`;
-  }),
-  EmbedLinkedDocBlockAdapterExtensions,
-  createBuiltinToolbarConfigExtension(flavour),
-  LinkedDocSlashMenuConfigExtension,
-].flat();
 
 export const EmbedLinkedDocViewExtensions: ExtensionType[] = [
   BlockViewExtension(flavour, model => {
