+++ blocksuite/affine/blocks/embed/src/embed-github-block/embed-github-spec.ts
@@ -5,7 +5,6 @@ import type { ExtensionType } from '@blocksuite/store';
 import { literal } from 'lit/static-html.js';
 
 import { createBuiltinToolbarConfigExtension } from '../configs/toolbar';
-import { EmbedGithubBlockAdapterExtensions } from './adapters/extension';
 import { embedGithubSlashMenuConfig } from './configs/slash-menu';
 import { EmbedGithubBlockInteraction } from './embed-edgeless-github-block';
 import { EmbedGithubBlockComponent } from './embed-github-block';
@@ -15,20 +14,6 @@ import {
 } from './embed-github-service';
 
 const flavour = EmbedGithubBlockSchema.model.flavour;
-
-export const EmbedGithubBlockSpec: ExtensionType[] = [
-  FlavourExtension(flavour),
-  EmbedGithubBlockService,
-  BlockViewExtension(flavour, model => {
-    return model.parent?.flavour === 'affine:surface'
-      ? literal`affine-embed-edgeless-github-block`
-      : literal`affine-embed-github-block`;
-  }),
-  EmbedGithubBlockAdapterExtensions,
-  EmbedGithubBlockOptionConfig,
-  createBuiltinToolbarConfigExtension(flavour, EmbedGithubBlockComponent),
-  SlashMenuConfigExtension(flavour, embedGithubSlashMenuConfig),
-].flat();
 
 export const EmbedGithubViewExtensions: ExtensionType[] = [
   FlavourExtension(flavour),
