+++ blocksuite/affine/blocks/embed/src/embed-html-block/embed-html-spec.ts
@@ -8,15 +8,6 @@ import { EmbedEdgelessHtmlBlockInteraction } from './embed-edgeless-html-block';
 
 const flavour = EmbedHtmlBlockSchema.model.flavour;
 
-export const EmbedHtmlBlockSpec: ExtensionType[] = [
-  BlockViewExtension(flavour, model => {
-    return model.parent?.flavour === 'affine:surface'
-      ? literal`affine-embed-edgeless-html-block`
-      : literal`affine-embed-html-block`;
-  }),
-  createBuiltinToolbarConfigExtension(flavour),
-].flat();
-
 export const EmbedHtmlViewExtensions: ExtensionType[] = [
   BlockViewExtension(flavour, model => {
     return model.parent?.flavour === 'affine:surface'
