+++ blocksuite/affine/blocks/embed/src/embed-iframe-block/embed-iframe-spec.ts
@@ -4,24 +4,11 @@ import { BlockViewExtension, FlavourExtension } from '@blocksuite/std';
 import type { ExtensionType } from '@blocksuite/store';
 import { literal } from 'lit/static-html.js';
 
-import { EmbedIframeBlockAdapterExtensions } from './adapters';
 import { embedIframeSlashMenuConfig } from './configs/slash-menu/slash-menu';
 import { createBuiltinToolbarConfigExtension } from './configs/toolbar';
 import { EmbedIframeInteraction } from './embed-edgeless-iframe-block';
 
 const flavour = EmbedIframeBlockSchema.model.flavour;
-
-export const EmbedIframeBlockSpec: ExtensionType[] = [
-  FlavourExtension(flavour),
-  BlockViewExtension(flavour, model => {
-    return model.parent?.flavour === 'affine:surface'
-      ? literal`affine-embed-edgeless-iframe-block`
-      : literal`affine-embed-iframe-block`;
-  }),
-  EmbedIframeBlockAdapterExtensions,
-  createBuiltinToolbarConfigExtension(flavour),
-  SlashMenuConfigExtension(flavour, embedIframeSlashMenuConfig),
-].flat();
 
 export const EmbedIframeViewExtensions: ExtensionType[] = [
   FlavourExtension(flavour),
