+++ blocksuite/affine/blocks/embed/src/embed-youtube-block/embed-youtube-spec.ts
@@ -5,7 +5,6 @@ import type { ExtensionType } from '@blocksuite/store';
 import { literal } from 'lit/static-html.js';
 
 import { createBuiltinToolbarConfigExtension } from '../configs/toolbar';
-import { EmbedYoutubeBlockAdapterExtensions } from './adapters/extension';
 import { embedYoutubeSlashMenuConfig } from './configs/slash-menu';
 import { EmbedYoutubeBlockInteraction } from './embed-edgeless-youtube-block';
 import { EmbedYoutubeBlockComponent } from './embed-youtube-block';
@@ -15,20 +14,6 @@ import {
 } from './embed-youtube-service';
 
 const flavour = EmbedYoutubeBlockSchema.model.flavour;
-
-export const EmbedYoutubeBlockSpec: ExtensionType[] = [
-  FlavourExtension(flavour),
-  EmbedYoutubeBlockService,
-  BlockViewExtension(flavour, model => {
-    return model.parent?.flavour === 'affine:surface'
-      ? literal`affine-embed-edgeless-youtube-block`
-      : literal`affine-embed-youtube-block`;
-  }),
-  EmbedYoutubeBlockAdapterExtensions,
-  EmbedYoutubeBlockOptionConfig,
-  createBuiltinToolbarConfigExtension(flavour, EmbedYoutubeBlockComponent),
-  SlashMenuConfigExtension('affine:embed-youtube', embedYoutubeSlashMenuConfig),
-].flat();
 
 export const EmbedYoutubeViewExtensions: ExtensionType[] = [
   FlavourExtension(flavour),
