+++ blocksuite/affine/blocks/embed/src/index.ts
@@ -1,22 +1,3 @@
-import type { ExtensionType } from '@blocksuite/store';
-
-import { EmbedFigmaBlockSpec } from './embed-figma-block';
-import { EmbedGithubBlockSpec } from './embed-github-block';
-import { EmbedHtmlBlockSpec } from './embed-html-block';
-import { EmbedIframeBlockSpec } from './embed-iframe-block';
-import { EmbedLoomBlockSpec } from './embed-loom-block';
-import { EmbedYoutubeBlockSpec } from './embed-youtube-block';
-
-export const EmbedExtensions: ExtensionType[] = [
-  // External embed blocks
-  EmbedFigmaBlockSpec,
-  EmbedGithubBlockSpec,
-  EmbedLoomBlockSpec,
-  EmbedYoutubeBlockSpec,
-  EmbedHtmlBlockSpec,
-  EmbedIframeBlockSpec,
-].flat();
-
 export { createEmbedBlockHtmlAdapterMatcher } from './common/adapters/html';
 export { createEmbedBlockMarkdownAdapterMatcher } from './common/adapters/markdown';
 export { createEmbedBlockPlainTextAdapterMatcher } from './common/adapters/plain-text';
