+++ blocksuite/affine/blocks/latex/src/index.ts
@@ -3,7 +3,6 @@ import type * as RemarkMathType from 'remark-math';
 export * from './adapters';
 export * from './commands';
 export * from './latex-block';
-export * from './latex-spec';
 
 // Global types
 declare type _GLOBAl = typeof RemarkMathType;
