+++ blocksuite/affine/blocks/list/src/index.ts
@@ -2,6 +2,5 @@ export * from './adapters/index.js';
 export * from './commands';
 export { correctNumberedListsOrderToPrev } from './commands/utils';
 export * from './list-block.js';
-export * from './list-spec.js';
 export * from './turbo/list-layout-handler';
 export * from './turbo/list-painter.worker';
