+++ blocksuite/affine/blocks/note/src/index.ts
@@ -5,6 +5,5 @@ export * from './config';
 export * from './edgeless-clipboard-config';
 export * from './note-block';
 export * from './note-edgeless-block';
-export * from './note-spec';
 export * from './turbo/note-layout-handler';
 export * from './turbo/note-painter.worker';
