+++ blocksuite/affine/blocks/paragraph/src/index.ts
@@ -2,6 +2,5 @@ export * from './adapters/index.js';
 export * from './commands';
 export * from './paragraph-block.js';
 export * from './paragraph-block-config.js';
-export * from './paragraph-spec.js';
 export * from './turbo/paragraph-layout-handler';
 export * from './turbo/paragraph-painter.worker';
