+++ blocksuite/affine/blocks/root/src/edgeless/utils/consts.ts
@@ -1,11 +1,3 @@
 export const DEFAULT_NOTE_CHILD_FLAVOUR = 'affine:paragraph';
 export const DEFAULT_NOTE_CHILD_TYPE = 'text';
 export const DEFAULT_NOTE_TIP = 'Text';
-
-export const FIT_TO_SCREEN_PADDING = 100;
-
-export const ATTACHED_DISTANCE = 20;
-
-export const SurfaceColor = '#6046FE';
-export const NoteColor = '#1E96EB';
-export const BlendColor = '#7D91FF';
