+++ blocksuite/affine/blocks/surface-ref/src/index.ts
@@ -1,4 +1,3 @@
 export * from './commands.js';
 export * from './surface-ref-block.js';
 export * from './surface-ref-block-edgeless.js';
-export * from './surface-ref-spec.js';
