+++ blocksuite/affine/blocks/table/src/index.ts
@@ -2,4 +2,3 @@ export * from './adapters';
 export * from './commands.js';
 export * from './selection-schema.js';
 export * from './table-data-manager.js';
-export * from './table-spec.js';
