--- blocksuite/affine/inlines/footnote/src/exts.ts
@@ -1,7 +1,0 @@
-import type { ExtensionType } from '@blocksuite/store';
-
-import { FootNoteInlineSpecExtension } from './inline-spec';
-
-export const inlineFootnoteExtensions: ExtensionType[] = [
-  FootNoteInlineSpecExtension,
-];
