+++ blocksuite/affine/inlines/footnote/src/index.ts
@@ -1,5 +1,4 @@
 export * from './adapters';
-export * from './exts';
 export * from './footnote-node/footnote-config.js';
 export { AffineFootnoteNode } from './footnote-node/footnote-node.js';
 export * from './inline-spec';
