--- blocksuite/affine/inlines/latex/src/exts.ts
@@ -1,11 +1,0 @@
-import {
-  LatexEditorUnitSpecExtension,
-  LatexInlineSpecExtension,
-} from './inline-spec';
-import { LatexEditorInlineManagerExtension } from './latex-node/latex-editor-menu';
-
-export const inlineLatexExtensions = [
-  LatexInlineSpecExtension,
-  LatexEditorUnitSpecExtension,
-  LatexEditorInlineManagerExtension,
-];
