--- blocksuite/affine/inlines/link/src/exts.ts
@@ -1,9 +1,0 @@
-import type { ExtensionType } from '@blocksuite/store';
-
-import { LinkInlineSpecExtension } from './inline-spec';
-import { linkToolbar } from './toolbar';
-
-export const inlineLinkExtensions: ExtensionType[] = [
-  LinkInlineSpecExtension,
-  linkToolbar,
-];
