+++ blocksuite/affine/inlines/link/src/index.ts
@@ -1,6 +1,5 @@
 export * from './adapters';
 export * from './command';
-export * from './exts';
 export * from './inline-spec';
 export * from './link-node';
 export * from './markdown';
