--- blocksuite/affine/inlines/mention/src/exts.ts
@@ -1,7 +1,0 @@
-import type { ExtensionType } from '@blocksuite/store';
-
-import { MentionInlineSpecExtension } from './inline-spec';
-
-export const inlineMentionExtensions: ExtensionType[] = [
-  MentionInlineSpecExtension,
-];
