+++ blocksuite/affine/inlines/mention/src/index.ts
@@ -1,4 +1,3 @@
 export * from './affine-mention';
 export * from './effects';
-export * from './exts';
 export * from './inline-spec';
