+++ blocksuite/affine/inlines/preset/src/index.ts
@@ -5,7 +5,6 @@ declare type _GLOBAL_ = typeof RichTextEffects;
 export * from './adapters/extensions';
 export * from './command';
 export * from './default-inline-manager';
-export * from './exts';
 export * from './inline-spec';
 export * from './keymap';
 export * from './markdown';
