+++ blocksuite/affine/inlines/reference/src/index.ts
@@ -1,5 +1,4 @@
 export * from './adapters';
-export * from './exts';
 export * from './inline-spec';
 export * from './reference-node';
 export * from './toolbar';
